[Fix] Revert accidental package renaming in integration tests.
