+++ jopa-integration-tests/src/test/java/cz/cvut/kbss/jopa/test/integration/AttributesTest.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library.
  */
-package cz.cvut.kbss.jopa.test.integration.virtuoso;
+package cz.cvut.kbss.jopa.test.integration;
 
 import cz.cvut.kbss.jopa.exceptions.AttributeModificationForbiddenException;
 import cz.cvut.kbss.jopa.test.OWLClassM;
