+++ jopa-integration-tests/src/test/java/cz/cvut/kbss/jopa/test/integration/environment/TestDataSource.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library.
  */
-package cz.cvut.kbss.jopa.test.integration.virtuoso.environment;
+package cz.cvut.kbss.jopa.test.integration.environment;
 
 import cz.cvut.kbss.ontodriver.Connection;
 import cz.cvut.kbss.ontodriver.DataSource;
